add an option to force compression and not just rely on the file suffix

+++ src/java/io/compgen/common/io/FileUtils.java
@@ -10,13 +10,25 @@ public class FileUtils {
 	public FileUtils() {
 	}
     public static OutputStream openOutputStream(String outputName) throws IOException {
-    	if (outputName == null || outputName.equals("-")) {
-            return System.out;
-        } else if (outputName.endsWith(".gz")) {
-            return new GZIPOutputStream(new FileOutputStream(outputName));
-        } else {
-            return new BufferedOutputStream(new FileOutputStream(outputName));
-        }
+    	return openOutputStream(outputName, false);
+    }
+
+    public static OutputStream openOutputStream(String outputName, boolean forceCompress) throws IOException {
+    	if (forceCompress) {
+        	if (outputName == null || outputName.equals("-")) {
+                return new GZIPOutputStream(System.out);
+            } else {
+                return new GZIPOutputStream(new FileOutputStream(outputName));
+            }
+    	} else {
+	    	if (outputName == null || outputName.equals("-")) {
+	            return System.out;
+	        } else if (outputName.endsWith(".gz")) {
+	            return new GZIPOutputStream(new FileOutputStream(outputName));
+	        } else {
+	            return new BufferedOutputStream(new FileOutputStream(outputName));
+	        }
+    	}
     }
 
 }
